apps/urmom/pyproject.toml
[project]
name = "urmom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pywin32>=311",
    "pyinstaller>=6.0.0",
    "watchdog>=4.0.0",
    "psutil>=7.2.1",
    "pyqt6>=6.10.2",
    "pillow>=12.1.0",
    "litellm>=1.80.17",
    "pystray>=0.19.5",
    "groq>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[dependency-groups]
dev = [
    "watchdog>=6.0.0",
    "litellm>=1.80.17",
]
